requirements.txt
fastapi
uvicorn
transformers
torch
youtube_dl
ffmpeg-python
openai-whisper
